requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-dotenv
jinja2
openai
tavily-python
wikipedia
requests
aisuite
docstring_parser

# Core
pdfminer.six
pymupdf           # optional but recommended for faster/better PDF text extraction
